Corrijo Makefile, modificando funciones de operaciones para que compilen correctamente

+++ includes/libft_push_swap/ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: samperez <samperez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/10 17:19:07 by samperez          #+#    #+#             */
-/*   Updated: 2024/10/10 18:51:07 by samperez         ###   ########.fr       */
+/*   Updated: 2024/12/03 12:10:39 by samperez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 // Measures the length of a string
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	int	i;
